langdeepagent/requirements.txt
# Core dependencies
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langgraph>=0.2.0

# DeepSeek and OpenAI
openai>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Async support
asyncio
aiohttp>=3.8.0

# Configuration and environment
python-dotenv>=1.0.0

# Data storage
sqlite3

# Web framework (for future API)
fastapi>=0.100.0
uvicorn>=0.23.0

# User interface (optional)
streamlit>=1.28.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Logging and monitoring
loguru>=0.7.0

# Utilities
python-dateutil>=2.8.0
pydantic>=2.0.0